Modal dialogs (which, in the TurboVNC Viewer, include Options and New Connection) block on the call to setVisible(true), so the call to setAlwaysOnTop(true) had no effect with those dialogs.  We don't want modal dialogs to be always on top anyhow, unless the viewer is running as an embedded applet.  In that case, the browser does not allow modal dialogs from the embedded applet to block the browser window (for security reasons), so clicking away from the dialog sends it to the back of the window stack.  For the casual user, this makes it appear as if the dialog has disappeared, even though technically it can be recovered.  Recovering it is cumbersome, however, since modal dialogs are not registered with the task switcher, and thus it is necessary to minimize all other windows in order to find the missing dialog.  The solution is to call setAlwaysOnTop(true) only for non-modal dialogs or if in embedded applet mode and to make the call prior to the call to setVisible(true).  This keeps all dialogs always on top when in embedded applet mode and only modeless dialogs (including the performance profiling dialog) always in top when in windowed mode.
This patch also removes the calls to setFocusable(), which have existed ever since the first incarnation of this viewer that was checked into the TigerVNC repository.  As with setAlwaysOnTop(true), the call to setFocusable(true) was actually not having any effect with modal dialogs, because it wasn't being executed prior to setVisible().  However, the call to setFocusable(false) was disabling the focusable state of all of the components in the dialog, which had the effect of breaking the ability to control it with the keyboard the next time it was popped up.


git-svn-id: svn+ssh://svn.code.sf.net/p/turbovnc/code/trunk@2684 799e4f7b-5fd2-41f6-823c-2ecc41bc7f0b