Fixed several issues related to the display of dialogs.  In applet mode, the ownerless dialogs (New Connection, Password, SSH, etc.) would disappear to the back of the window stack if the user clicked away from the dialog, and they could not be recovered with Alt-Tab.  Further, under Java 7 and Linux, that same thing would happen in standalone mode.  Normally, if a JDialog is ownerless, it is assigned to a "shared frame", and the shared frame is registered with the task switcher, even though it has no taskbar icon.  Such is not the case with Java 7 and Linux, however.
For the dialogs that we control, we now force the dialogs to have a null owner (i.e. not assigned to the shared frame), which ensures that Java 6 and later will add them to both the taskbar and task switcher on Linux and Windows (NOTE: you can't assign a null owner to a JDialog in Java 1.5 and earlier, but since Java 1.5 already adds JDialogs to the task switcher, it doesn't matter.)  A null owner can't be assigned to a JOptionPane, so for those dialogs (which are meant to be short-lived anyhow), we make them always on top.  In applet mode on Linux and Windows, the ownerless dialogs are now registered with the taskbar and task switcher, but not so on Mac, so for that reason, ownerless dialogs are made always on top when in applet mode (NOTE: when running as an embedded applet, all dialogs are ownerless and are thus always on top.)

This patch refactors the Dialog class somewhat so that the owner of the dialog is explicitly specified in all cases.  This allows us to assign a null owner, per above, but it also allows us to assign the proper owner to the JDialog when displaying the Options or Clipboard dialogs.  This fixes various cosmetic issues, such as:

-- When the Options dialog was invoked from the New Connection dialog, it was not behaving entirely like a modal dialog.  It was possible to click on the New Connection dialog and bring it to the front.

-- In full-screen mode on Linux/Gnome, popping up the Options or Clipboard dialogs would cause the viewer to temporarily exit full-screen mode.


git-svn-id: svn+ssh://svn.code.sf.net/p/turbovnc/code/trunk@2705 799e4f7b-5fd2-41f6-823c-2ecc41bc7f0b