Make sure we actually shut down the listener when "Close listener" is selected in the tray menu.  There is some subtlety to this.  One would think that simply closing the socket would abort the current accept() operation, but it doesn't on all platforms (I'm looking at you, Windows.)  It is necessary to interrupt the thread to actually get the listener to abort.  Also on Windows, this causes the viewer to still pretend like it's accepting connections, so from the point of view of the server, vncconnect fails silently until the viewer process exits.  Not sure how to fix that (it doesn't occur on other platforms.)
With the Windows Firewall enabled, even more bizarre behavior can ensue, whereby vncconnect will hang when trying to connect to the viewer after the listener has been closed, but then, upon relaunching the viewer, the connection will go through.

This patch also removes the "Close listener" option from the tray menu if "NoNewConn" is enabled, because closing the listener in that case is the same as closing the viewer.  Further, it makes any errors encountered when trying to set the tray icon non-fatal, and it removes a debug print statement from one of the error handlers in TrayMenu.


git-svn-id: svn+ssh://svn.code.sf.net/p/turbovnc/code/trunk@2756 799e4f7b-5fd2-41f6-823c-2ecc41bc7f0b