Expose CL 2 and interframe comparison in the GUI and redefine the encoding method presets such that CL 2 is used along with Low-Quality JPEG (makes sense-- that's supposed to be a minimum-bandwidth-usage mode) and interframe comparison is used along with Medium- and Low-quality JPEG and Lossless Tight + Zlib.  Because of this, we can no longer use the "Zlib compression level" label in the GUI, because CL 2 has a different palette threshold from CL 1.  We simply refer the user to the docs for an explanation of what "compression level" means.
Another ramification of this patch is that we can no longer automatically enable the TightVNC/TigerVNC compatibility GUI whenever CL>=1, because the user might be setting CL=5-7 in order to activate the ICE-- and this is now a completely legal setting in the GUI.  Thus, we introduce a new "CompatibleGUI" parameter that enables the compatibility GUI.  Further, when the compatibility GUI is enabled, we no longer send the TurboVNC Fine quality level/subsampling pseudo-encodings.  We send a Tight quality pseudo-encoding (0-9) only.


git-svn-id: svn+ssh://svn.code.sf.net/p/turbovnc/code/trunk@2871 799e4f7b-5fd2-41f6-823c-2ecc41bc7f0b