Java: Major overhaul of SSLEngineManager The SSLEngineManager was basically an abomination.  The work is now done as it should be, with the buffers being fed and the engine deciding when data is consumed or produced.  The engine should be much more robust now as well.
Additionally, although JRE 7 supports the TLS v1.1 and TLS v1.2
protocols, they are not actually enabled by default.  The JSSE reference
cites compatibility reasons for this, but this doesn't appear to be the
case with the TurboVNC Server.  Furthermore, TLS v1.1+ will be enabled
by default in JRE 8.

The regular expression for enabling anonymous DH cipher suites was too
narrow and excluded the elliptic curve ciphers, which are now ordered
ahead of the ephemeral ciphers by the default security provider.

This incorporates the following TigerVNC commits:
https://github.com/TigerVNC/tigervnc/commit/e20370827bf9441c91de0baeb696a4cf9d6f75a6
https://github.com/TigerVNC/tigervnc/commit/93721e78f145f1e7d92cd8c5a1f8e0b0b502c6ce