Fix two regressions caused by previous automatic desktop resizing commits:
-- When not using Lion FS mode, the full-screen window did not fill the display unless the remote desktop size was greater than or equal to the span size.  This got confusing, because background windows were visible around the edges.

-- When the viewer was started in full-screen mode but the remote desktop size was less than the span size, the remote desktop was not automatically resized.

This fix has one unfortunate side effect, which is that, if the remote desktop is smaller than the full-screen viewport, it won't be centered as long as automatic desktop resizing is enabled.  There is unfortunately no easy way around that.  viewport.dx/dy are supposed to address it, but the logic in componentResized() that sets those variables doesn't work properly with automatic desktop resizing.  The case of the remote desktop being smaller than the viewport is rare if automatic desktop resizing is enabled.  It will only occur if multiple viewers are connected, one of them is in full-screen mode, and other viewers cannot resize to accommodate that resolution (per the previous commit message, non-full-screen viewers will always win that fight.)


git-svn-id: svn+ssh://svn.code.sf.net/p/turbovnc/code/trunk@2975 799e4f7b-5fd2-41f6-823c-2ecc41bc7f0b