Fix several cosmetic issues in the Java blitter:
(1) When Swing double buffering was enabled, an animation or video was playing on the server, and the user was interacting with one of the dialogs (Options, etc.) on the client, moving the dialog or interacting with its components would cause pixels from the remote desktop viewport underlying the dialog to be displayed at the upper left of the viewport.  This was known to be an issue with Java 7 on Windows but may have affected other platforms.

(2) Jumping the scrollbars from top to bottom on OS X under Java 8 would sometimes cause regions from the top or bottom of the remote desktop viewport to remain if it was scrolled to the opposite end.

(3) White pixels would sometimes appear around the Options Dialog GUI components on OS X when interacting with them.

(4) When using full-screen mode on Linux/Windows or non-Lion full-screen mode on Mac, scrolling the viewport would sometimes leave junk pixels in the areas of the window not covered by the remote desktop image.

The cause of these issues was two-fold:

-- We really need to be painting only in the EDT, period.  Not doing so was the primary cause of (1) and (2) and possibly (3).  However, using SwingUtilities.invokeAndWait() has too much overhead for small updates.  We now use repaint() instead of paintImmediately(), primarily because it's a fast way of ensuring that the painting occurs in the EDT.  There are other advantages, such as the fact that Swing can coalesce rapid back-to-back updates into a single paint.  We still use paintImmediately() when benchmarking, since concurrency isn't an issue there and we want to avoid the afore-mentioned update coalescing.  The other viewers take similar steps to synchronously blit whenever benchmarking is enabled.

-- With double buffering disabled, we can't call super.paintComponent(), nor can we call setOpaque(false), because both cause flickering when the RFB thread triggers a repaint.  However, not calling super.paintComponent() causes artifacts such as (4) above (and possibly (3) as well.)  Previously, we used a hackish "repaintBackground" mechanism, whereby super.paintComponent() was called the first time after the viewport was resized, but that was insufficient.  We now detect in paintComponent() whether double buffering is enabled (which means that the repaint was system-triggered, not application-triggered, and we call super.paintComponent() only in that case.


git-svn-id: svn+ssh://svn.code.sf.net/p/turbovnc/code/branches/1.2.x@3106 799e4f7b-5fd2-41f6-823c-2ecc41bc7f0b