Java: Fix NullPtrExc when X.509 cert unspecified
When attempting to connect using one of the X.509 security types without
specifying a CA certificate using the X509CA parameter, the viewer would
fail with a NullPointerException in CSecurityTLS.java:104.  This patch
does the following:

-- Adds an automatic certificate saving feature such as the one in the
TigerVNC FLTK viewer.  This pops up a confirmation dialog if the CA
certificate is not trusted and, if the user answers "Yes", it stores
the certificate in a certificate store so it will be trusted for
subsequent connections.
-- If the server is using a certificate from a top-level authority, the
viewer now pulls in CA certificates from all trusted authorities known
to the JRE.
-- Allows the Java viewer to read CA certificate files containing
multiple certificates concatenated together.

Incorporates the following TigerVNC commits:
https://github.com/TigerVNC/tigervnc/commit/95f39a5685bbf389200e0c7abea5d7db50755572
https://github.com/TigerVNC/tigervnc/commit/3ee966ed4e2334368c40f4f3e28de1d1c6bb4159
https://github.com/TigerVNC/tigervnc/commit/395a93418b6acdb6a02ab575c540152dc81295c5