Server: Fix TLS incompatibility w/ TigerVNC Viewer The 512 argument to DH_generate_parameters_ex() was the result of futzing around in an attempt to fix a long delay that occurred when either the TigerVNC Viewer or our Java viewer connected with Anonymous TLS encryption.  Although setting that argument to 512 did work around the performance issue, it also set the DH prime length to 512 bits, which wasn't secure (and it caused the TigerVNC FLTK Viewer to abort with an error complaining that our prime wasn't long enough.)
It turns out that DH_generate_parameters_ex() isn't really intended to
be called in real time.  It is used for generating "strong" primes that
can be reused by the same server over and over again (in the past,
generating these primes took minutes, so the OpenSSL developers
recommended generating the prime at install time.  I guess I shouldn't
complain that it was only taking 15 seconds on my machine.)  One can
generate a prime much more quickly from a DSA.  This prime is considered
"weak", but since a new one is generated for each client connection, it
should be reasonable for use with anonymous TLS (which itself isn't
immune from all attacks.)  Anyhow, this method is the same one GnuTLS
uses (AFAICT.)

Note that the OpenSSL documentation recommends also setting the
SSL_OP_SINGLE_DH_USE option when using a prime generated from a DSA,
but this assumes a server that sets up the SSL context once and calls
SSL_accept() in a loop.  Since we're setting up a new SSL context for
each connection, it doesn't matter whether we use SSL_OP_SINGLE_DH_USE
(which causes OpenSSL to generate the DH key during negotiation) or
whether we pre-generate the key as we're currently doing.  Either way,
we get a new key for every connection.