Server: Fix build if X11 include dir is non-std
nvctrlext.c includes Xlib.h, which is meant to be provided by the system
(since we're using the system's version of libX11 to communicate with
the 3D X server.)  Thus, we need to add the system's X11 include path
to the set of include directories.  This specifically fixes the build
under FreeBSD, which keeps its X11 headers under /usr/local.