Java: X.509 hostname verification
This adds hostname verification for X.509 certificates using either
the subjectAltName (SAN) X.509 v3 extension or, if SAN is not present,
the CN subject field.

If hostname verification fails, a dialog is displayed, and the user has
the option of aborting or continuing the connection.

This also cleans up exception handling in CSecurityTLS.java and removes
some vestigial code.

Incorporates the following TigerVNC commits:
https://github.com/TigerVNC/tigervnc/commit/218be2803a981e7449f182ee8657e225ef52f3c0
https://github.com/TigerVNC/tigervnc/commit/63ee86708feda53eebcd19db97ed6c5da47cb2b5
https://github.com/TigerVNC/tigervnc/commit/fdb669952da1ca60a88291632de7834b8f7ed8dc