Java: Fix CurrentMonitorIsPrimary issue on Windows
For some reason, Java on Windows (unknown if this affects all versions)
will reverse the screen order that is reported to the application, so
the right-most screen is reported first and the left-most screen is
reported second.  The Java TurboVNC Viewer handled this previously by
always treating the screen with the left-most and top-most coordinates
as the primary, even if it wasn't reported as Screen 0.  However, this
messed up the CurrentMonitorIsPrimary feature.  If the viewer window was
mostly on the right screen, then that feature would set the right screen
(Screen 0) as the primary during the first iteration of the loop, then
it would set the left screen (Screen 1) as the primary during the second
iteration, since Screen 1's coordinates were to the left or top of
Screen 0's.  This patch prevents the primary from being set based on
coordinates after it has already been set based on the window position.

Hopefully this really fixes #11.