Java: Allow specifying max. clipboard xfer size
This incorporates the following TigerVNC commits:
https://github.com/TigerVNC/tigervnc/commit/fefa48d3c47a2cfbedf6dd544354a2db97760b85
https://github.com/TigerVNC/tigervnc/commit/d121c1430dd55da0a1ca6160606cc180946f2401

which purported to fix an issue whereby too much data on the local
clipboard could cause the viewer's heap size to expand to an enormous
size and exceed the JVM's maximum heap size, thus causing the viewer to
abort.  I have never been able to reproduce this, however.  Both with
and without this patch, the viewer will frequently generate OutOfMemory
errors when transferring a huge amount of data via the clipboard, but
extensive testing has never revealed a case in which the data
subsequently failed to transfer after the OOM error was generated.
Furthermore, the clipboard transfer overhaul in TigerVNC regresses
behavior in other ways, per my testing, and it still generates the same
OOM errors.

Thus, the main purpose of this patch is to provide a means of limiting
the size of the clipboard data that the viewer tries to transfer, so as
to avoid exceeding the server's clipboard transfer size limit.