Java: Avoid unnecessary scrollbars
If the remote desktop width was too large to be accommodated without
scrollbars but the remote desktop height wasn't, or vice versa, then
the viewer would display two scrollbars, when actually only one was
needed.  This commit changes the window size calculations for "server"
(manual) remote desktop sizing such that this case is properly handled.

Fixes #13