Unset DBUS_SESSION_BUS_ADDRESS in default xstartup
This turns out to be necessary when starting the TurboVNC Server from
within an existing GNOME 2 or MATE session.  Otherwise, the existing
D-Bus daemon will have set the DBUS_SESSION_BUS_ADDRESS environment
variable, which causes the new instance of GNOME running in TurboVNC
to attempt to share the existing D-Bus instance (resulting in the
infamous "could not acquire name on session bus" error.)

https://bugzilla.redhat.com/show_bug.cgi?id=497592

This patch has been in TigerVNC since almost the beginning of that
project, but since the TurboVNC Server is typically started from web
portals and SSH sessions (neither of which are affected by this issue),
it was not understood until recently how to reproduce the issue.