Server: Optionally use static XI devs for tablet
In some cases, it may be desirable to launch a specific graphics
application rather than a window manager from xstartup.turbovnc.  Since
most graphics applications check for the existence of extended input
devices at startup, this situation broke the extended input device
feature in TurboVNC-- because that feature, as implemented, didn't
create devices on the server until a viewer connected and requested the
creation of the devices.

This patch adds an Xvnc command-line argument (-virtualtablet) that
changes the behavior of the extended input device feature such that it
assumes that all extended input events will be coming from a Wacom or
similar tablet.  With this assumption, the server can statically create
a virtual stylus and eraser device and map the tablet events received
from the viewer to those static devices.

This option is generally less flexible than the more generic remote X
Input option, since the virtual tablet option can only accommodate
tablets.  The virtual tablet option also assumes that the eraser device
from the tablet will be named "*eraser*" (case-insensitive), so it maps
all extended input events from that device to the virtual eraser and all
other extended input events to the virtual stylus.  Furthermore, this
option assumes that the valuators from the extended input device will
correspond to Abs X, Abs Y, Abs Pressure, Abs Tilt X, Abs Tilt Y, and
Abs Wheel, in that order.  It is unknown whether that is generally the
case for all tablets or even for all Wacom models.