Fix several issues with ext. input device feature
Whereas GIMP and Krita will ignore core pointer events whenever an
extended pointer device is being used, other applications such as Nuke
will not, and the fact that the TurboVNC Viewer was sending a "regular"
pointer event along with each extended pointer event was causing
problems with those applications.  This patch attempts to fix the issue
by eliminating the X Input listener thread and instead processing X
Input events within the EDT-- and, more specifically, only processing
the specific type of event that Java was notified about.  This allows us
to figure out whether the Java event was triggered by an X Input event,
and if it was, we send the event using GII rather than the standard RFB
pointer message.

It is hoped that this patch will also address a reported interaction lag
with extended pointer events, which could have been caused by heap bloat
in the JVM (a new object was previously allocated for each new event,
whereas now the same event object is reused) or possibly just thread
contention between the EDT and the X Input listener thread.