Server: Fix regression caused by libXfont upgrade
Whereas libXfont-1.4.5 assumed that the readlink() function was always
available, libXfont-1.4.9 introduced a new macro (HAVE_READLINK) to
allow for building on systems without readlink().  If this macro is
undefined, then font catalogue support is not built into libXfont, and
this was causing several font-related issues when running TurboVNC 2.1
alpha on modern Linux releases (such as RHEL 6+.)  Since all of the
platforms we support have readlink(), the easiest solution was just to
define that macro for the whole libXfont build.

Fixes #24