Viewer: always show compression level in titlebar
In TurboVNC releases prior to 2.0.x, it made sense to only show the
compression level when it was > 1 (because none of the preset encoding
methods used a compression level > 1.)  However, since 2.0.x, the preset
encoding methods have used different compression levels (1 for PL JPEG
and LT + Zlib, 6 for MQ JPEG, 7 for LQ JPEG, 0 for LT), but the titlebar
logic was never updated to accommodate this.  Instead of attempting to
detect the preset and display that in the titlebar, it makes more sense
to just print the compression level that is in use.