Viewer: dead key support
Dead keys and AltGr-modified symbols should now work properly with the
Windows (native) TurboVNC Viewer.  Dead key support in the Java viewer
is still spotty for the following reasons:

* Java 6 doesn't use the proper key code when a "modified" dead key is
pressed (such as Shift-accent to produce a dead grave accent on a
Swedish keyboard, or Shift-accent to produce a dead tilde on a U.S.
International keyboard.)  Instead, the key code of the unmodified dead
key is communicated to the application.  No way to work around this.
Use Java 8.

* On Windows (under Java 8), Java uses the unmodified dead key code for
the 2nd and subsequent even-numbered presses/releases of a modified dead
key.  Thus, pressing a modified dead key twice doesn't produce the
expected behavior.  For instance, pressing dead tilde twice should
produce a tilde, but instead it does nothing.  No easy way to work
around this in the code, but pressing the modified dead key + space
works properly as an end-user workaround, since that also produces the
"live" key symbol corresponding to the dead key.

* On OS X, Java doesn't generally communicate the same key code when a
dead key is released as it does when the key is pressed.  On Linux, Java
doesn't generally communicate a KeyPress event for dead keys at all (it
only communicates a KeyRelease event.)  No known way to work around
these issues at the moment.