Build: Move display of C/LD flags (again)
win/CMakeLists.txt makes some modifications to the compiler flags, so it
makes the most sense to display the compiler/linker flags from within
the win/ and unix/ directories.  This captures any such changes to the
flags made within those subdirectories and also guarantees that we're
using the C/C++ compilers.