Server: Add -poll option to tvncconfig
Referring to #20, some users have experienced issues with Windows
clients whereby clipboard synchronization will work upon initially
connecting to the server but will stop working after a while.

TigerVNC users have reported this problem as well, and running vncconfig
with -poll 1000 is one known way to work around the issue.

Eventually, it would be nice to properly diagnose and solve the problem,
but it is difficult to reproduce.