Server: Fix offscr. win interference w/ comp. WMs
The easiest way to reproduce the interference is described in
TigerVNC/tigervnc#211.  Under Gnome 3 (I tested with RHEL 7, but this
likely affected other platforms and WM's as well), open an xterm on
Workspace 1 and type "ping {some website}".  Switch to Workspace 2 and
open another xterm.  When the xterm on Workspace 1 starts automatically
scrolling, it interferes with the rendering of the xterm on Workspace 2.

The underlying cause is that the visibility tests in draw.c were not
stringent enough.  They checked whether the destination drawable (in the
case of rfbGlyphs() and rfbComposite()) or the source drawable (in the
case of rfbCopyArea()) was a window and was viewable, but there is one
other relevant case:

* The drawable is a window and is viewable, but it is off-screen (that
is, its pixmap is not the same as the screen pixmap), in which case it
should be treated as invisible from the POV of VNC.

Based on:
https://github.com/TigerVNC/tigervnc/commit/9986dce0ae20b0821a5e189f58e86a764079048b