Server: Cancel deferred updates on screen resize
Referring to #15, when the screen saver in Linux Mint activates while
there are no viewers connected to the TurboVNC server, then the server
will sometimes stop sending updates the next time a viewer connects.
Specifically, this is triggered whenever the viewer requests a new
desktop size (which the TurboVNC Viewer will do upon connecting, if
automatic desktop resizing is enabled.)  For unknown reasons, the
reproducibility of this is either close to 100% or 0.  When the server
is in the right mood, then the issue can be reproduced by disconnecting
all viewers, resizing the remote desktop:

    DISPLAY=:1 xrandr -s 1600x1200

then starting the screensaver on it:

    DISPLAY=:1 mate-screensaver-command -a

then connecting with the TurboVNC Viewer with ADR enabled, thus
triggering a remote desktop resize request (assuming that the
client-side window is not 1600x1200.)

For reasons I can't fully explain, the issue is somehow caused by a
deferred update that is triggered prior to the screen resize but not
actually sent until after the resize.  It also seems to be related to
how the screen saver grabs the display, because the server doesn't
completely lock up-- it just stops sending updates.  Although the
condition is almost always reproducible when the server is in the right
mood, it appears to be somewhat racy and heisenbuggy, so inserting too
many print statements into the code makes it go away.

At any rate, cancelling pending FBUs when the screen is resized is a
good idea, irrespective of this specific bug.

Closes #15.