Java: Fix "Class not found" error (TJException)
libjpeg-turbo 1.5 uses a new class (TJException) to report errors from
the underlying TurboJPEG C library, but our build of VncViewer.jar was
not including that class.  This caused the Java viewer to produce the
following run-time error:

TightDecoder: WARNING: Could not initialize libjpeg-turbo:
TightDecoder:   Class not found: org/libjpegturbo/turbojpeg/TJException
TightDecoder:   Using unaccelerated JPEG decompressor.

and subsequently fail to accelerate JPEG decompression.

It's better to go ahead and include all of the classes in
org/libjpegturbo/turbojpeg, even though we only use the decompressor, in
order to prevent such issues from occurring in the future.  This adds
only a trivial amount of overhead to the JAR.