ChangeLog: Document removal of vncserver -no* args
In TurboVNC 2.0.x, the default was for vncserver to enable all of the
permitted authentication types in Xvnc by passing the corresponding
arguments (-rfbauth, -otpauth, -pamauth) to Xvnc, and the -no* arguments
to vncserver were used to disable specific auth methods for a particular
TurboVNC session, even if those auth methods were permitted on the
system.

In TurboVNC 2.1, there's a new argument (-securitytypes) that controls
which security types are enabled (a "security type" is a combination of
an authentication type and an encryption type.)  The default in Xvnc is
effectively "-securitytypes TLSVnc,TLSOtp,TLSPlain,X509Vnc,X509Otp,
X509Plain,VNC,OTP,UnixLogin,Plain", but this can be overridden in order
to disable some of the default security types or to enable any of the
"None" security types ("None", "TLSNone", or "X509None"), all of which
require no authentication (NOTE: -securitytypes is passed through from
vncserver to Xvnc, so refer to the Xvnc man page for a description of
that parameter.)

There isn't really a literal equivalent of the old -no* options.  I
toyed with the idea of keeping the -no* options in vncserver and making
them wrappers for -securitytypes, but this raised additional questions,
such as whether those options should work literally the same as they did
in TurboVNC 2.0.x (i.e. whether using them would imply disabling
encryption.)  Also, the problem with implementing backward-compatible
-no*  options in vncserver is that the -securitytypes argument can be
used not just to override the default security types that are available
but to override the default order as well.  In TurboVNC 2.0.x, the
preferred order of authentication types could be specified in the global
security config file (/etc/turbovncserver-security.conf), and specifying
-noauth or -novncauth, etc., didn't affect this order.  There isn't any
easy way to implement an equivalent behavior using the -securitytypes
argument, so I felt it best to force the issue by removing the -no*
arguments in vncserver and forcing users to switch to -securitytypes,
which is a more standard way of doing this sort of thing.

Addresses a concern raised in #39