Fix/simplify setting of CMake build timestamp vars
If wmic.exe wasn't available, then CMakeLists.txt would call
"cmd /C date /T" and parse the result in order to set the BUILD
variable.  However, the parser assumed that the date was in MM/DD/YYYY
format, which is not generally the case unless the user's locale is U.S.
English with the default region/language settings for that locale.

This commit modifies CMakeLists.txt such that it uses the
string(TIMESTAMP) function available in CMake 2.8.11 and later to set
the BUILD variable, thus eliminating the need to use wmic.exe or any
other platform-specific hack.

Furthermore, this commit also uses string(TIMESTAMP) to set JAVA_DATE
and JAVA_TIME, which eliminates a similar (but non-fatal) parsing issue
when using non-U.S. locales on Windows machines.  The JAR timestamp is
now stored using the ISO 8601 date/time formats.