Try to actually open unix sockets
It's probably more reliable and direct to just try and actually open the
needed unix sockets instead of testing for existence of the associated
paths and lines in /proc/net/unix.

There is some ancient-seeming stuff in the code, using pack() directly
instead of pack_sockaddr_in() but for the unix socket addresses I used
pack_sockaddr_un().  If Perl4 compatability really is needed, it's
possible to use something like

    pack('S a108', $AF_UNIX, "/tmp/.X11-unix/X$n")
    pack('S a*', $AF_UNIX, "\0/tmp/.X11-unix/X$n")