Server: Fix inf. ping-pong w/ -cursorshape -nocu
This fixes a regression introduced in 2.0 by migrating the server to
the X.org 7.7 code base.

When cursor shape updates (client-side cursor rendering) were enabled
and continuous updates were disabled, changing the cursor shape would
trigger an infinite ping-pong between viewer and server, caused by the
following sequence of events:

1. The server receives a framebuffer update request (FBUR.)
2. In response to the FBUR, the server checks whether it needs to send
   a new framebuffer update (FBU.)  Because the cursor shape has
   changed, a new FBU is triggered, thus causing
   rfbSendFramebufferUpdate() to be called.
3. rfbSendFramebufferUpdate() notices that the cursor has been drawn to
   the framebuffer, so it calls rfbRemoveCursorAllDev() to remove it
   from the framebuffer (in order to avoid cursor duplication, since the
   cursor is being drawn on the client.)
4. rfbRemoveCursorDev() triggers a CopyArea operation that adds to the
   modified region.
5. The cursor shape FBU is sent.
6. The sprite block handler, which runs when the server is idle, notices
   that the cursor isn't drawn, so it calls rfbRestoreCursor() to
   re-draw it.
7. rfbRestoreCursor() triggers a PushPixels operation that
   adds to the modified region.
8. The viewer sends back a new FBUR in response to the FBU sent in
   Step 5.
9. Because the modified region has changed, the new FBUR immediately
   triggers a new FBU, so the process repeats from Step 2.

The cycle is only broken if a non-cursor-related FBU is triggered on the
server.  A non-cursor-related FBU will cause rfbSendFramebufferUpdate()
to complete without calling rfbRemoveCursorAllDev(), so the next FBUR
from the viewer will not immediately trigger a new FBU.

The solution was simply to avoid adding to the modified region if
dontSendFramebufferUpdate is TRUE (which it will be in the body of
rfbSpriteRemoveCursor() and rfbSpriteRestoreCursor().)  This is similar
to the approach taken by TigerVNC (except that they don't use a
sprite handler at all.)  The TigerVNC logic also avoids adding to the
copy region while the cursor is being drawn or removed, and this commit
does likewise, although it's unknown if that is really necessary.

Fixes #40