Server: (mostly) fix server-side cursor rendering
Migrating to the X.org 7.7 code base introduced a completely new sprite
handler, and since the server-side cursor rendering code in TurboVNC is
based on that sprite handler, the X.org upgrade broke server-side cursor
rendering, causing severe rendering artifacts.  This bug has gone
unnoticed or unreported until now (server-side cursor rendering is not a
widely used feature.)

This commit makes our server-side cursor rendering implementation work
as well as the X.org code will allow it to work, but there are still
minor issues, such as the pointer flickering when a window is dragged,
that seem to be endemic to any software cursor rendering implementation
that is based on a recent X.org release:

https://bugs.launchpad.net/ubuntu/+source/xserver-xorg-video-fbdev/+bug/1034470

Fixing these remaining issues properly will require abandoning the
sprite code altogether and adopting a server-side cursor rendering
mechanism similar to that of TigerVNC (they don't actually render the
cursor on the remote framebuffer at all.  They simply store it as an
image whenever the shape changes, and they send the image as a
separate framebuffer update to any clients that need it.)