Java: Send new desktop size requests immediately
This commit causes the Java viewer to request a full framebuffer update
immediately after the Options dialog is dismissed, if the remote desktop
size was changed in that dialog and the new desktop resize mode is
"Auto" or "Manual".  Previously, the new desktop size was not
communicated to the server until the next framebuffer update was
received.

This also avoids an unnecessary viewport reconfigure when switching the
remote desktop size mode to "Server".