Server: Create per-session XDG run-time directory
Normally, systemd creates a separate XDG session each time a user logs
in (through SSH or GDM or whatnot.)  Each XDG session has its own ID
that is unique within the system, but all XDG sessions share the same
run-time directory.  This basically means that the TurboVNC session, if
left to its own devices, will inherit the state of any other GNOME 3
session on the system, which isn't necessarily what we want.  In
particular, #51 is a case in which this persistent state interferes with
the correct operation of TurboVNC (because the screensaver seems to
remember that it was activated in a previous TurboVNC session, thus
causing it to immediately activate in the new TurboVNC session.)  Some
have suggested clearing the XDG_SESSION_ID environment variable, but
that creates other issues (specifically, in my testing, it can cause the
screensaver to spin forever when attempting to unlock the screen.)

This commit causes vncserver to create a per-session XDG run-time
directory, assign the XDG_RUNTIME_DIR environment variable to that
per-session directory, and remove the directory when the TurboVNC
session is killed.  This might technically be an abuse of the XDG Base
Directory Specification
(https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html):

"The lifetime of the directory MUST be bound to the user being logged
in.  It MUST be created when the user first logs in and if the user
fully logs out the directory MUST be removed.  If the user logs in more
than once he should get pointed to the same directory, and it is
mandatory that the directory continues to exist from his first login to
his last logout on the system, and not removed in between."

However, in practice, GNOME 3 does not seem to be designed such that
multiple instances of it can share the same XDG run-time directory
without collision, so I'm not sure how else to work around this.  It
seems that, if the user's primary XDG run-time directory is used to
store temporary files created within the TurboVNC session, then some
state from the TurboVNC session can either conflict with the state from
other running TurboVNC sessions, or it can persist past the termination
of the TurboVNC session.

Fixes #51