Mac: Make "OracleJava" the preferred/default DMG
Further research has revealed that the Quartz-accelerated Java 2D
blitter in Java for OS X (Apple Java) is disabled not just on newer Mac
hardware but on all Macs running Yosemite and later.  This has led to
confusion as users upgraded their older Macs and discovered that the
TurboVNC "AppleJava" package suddenly performed poorly or produced
visual artifacts.

This commit removes the "OracleJava" suffix from the Oracle Java-based
TurboVNC package, to emphasize that it should be used more often than
not.  This commit also makes that package the default when executing
'make dmg' (the AppleJava package is now built with 'make compatdmg'),
and it lays the groundwork for removing support for Apple Java in a
future release of TurboVNC.

The documentation has also been updated to reflect the new findings.