Build: Improve TurboJPEG detection
This synchronizes the build system behavior with VirtualGL and
provides the following improvements relative to
a34ecf2a4b8a1bfc19ba6e4703865574f8ee7974:

- The build system no longer changes DEFAULT_TJPEG_INCLUDE_DIR if
TJPEG_INCLUDE_DIR is overridden, so the documentation for
TJPEG_INCLUDE_DIR will still show the correct default.

- The build system no longer tries to fall back to the system's
version of TurboJPEG if TJPEG_LIBRARY is overridden.  This provides a
way to sanity-check a particular value of TJPEG_LIBRARY.