init.d: Don't require ~/.vnc/passwd to be present
If the arguments to vncserver, or the permitted security types in the
security configuration file, only specify security types that don't
require a VNC password file, then there is no need for the init.d script
to check for the presence of that file.  That behavior was a relic of
TightVNC.  This commit removes that check from the init.d script and
instead causes the script to pass a new argument (-quiet) to vncserver.
This argument will cause vncserver to fail if the permitted and selected
security types require a VNC password file but none exists.