Build: Fix Java build dependencies
Since the JAR build command has several steps, it is necessary to use
an intermediate filename (VncViewer-unsigned.jar) for the unsigned JAR
file in order to ensure that the command will be re-run if the JAR
signing step fails.  This commit also causes the JAR build command to be
re-run if the TurboJPEG JAR changes on disk.