Server: Implement -disconnect option
... to optionally emulate the behavior of other VNC flavors and X
proxies, which disconnect existing users if a new viewer successfully
connects/authenticates without the "shared session" option enabled.

NOTE: Implementing this feature revealed a bug in the server's TLS
implementation whereby forcibly shutting down the SSL socket from the
server side while the Java viewer was trying to read from it would cause
the viewer to get into an infinite read loop.  It seems that it is
necessary to invoke shutdown() on the server prior to destroying the SSL
context, in order to ensure that any pending SSL writes are flushed.